Let `ProjectAccessorClassPathTest` print the emitted Kotlin metadata
To make debugging differences between the emitted bytecode and the emitted
source fragments easier.
